requirements.txt
flask
flask-login
flask-markdown
Flask-WTF
pyapi-gitlab
itsdangerous
markdown2